backend/requirements.txt
groq
python-dotenv
speechrecognition
pyaudio
pocketsphinx
pyttsx3
wikipedia
pycaw
comtypes
